+++ demo_examples/Chile_GRAV_4_Miller/LdM_grav_obs.grv
@@ -1,0 +1,192 @@
+191				
+362472.4	6008599.5	2185.513	-11.5534	0.05
+363047.4	6007741.5	2182.354	-14.4465	0.05
+365885.1	6010286.9	2181.814	-5.3319	0.05
+363446.6	6005839.3	2190.788	-16.3014	0.05
+364134	6005156.7	2173.695	-13.0721	0.05
+364417	6003660	2189.776	-0.1396	0.05
+367344.5	6005950.4	2194.062	5.6872	0.05
+360575.1	6011701.8	2267.91	-0.7109	0.05
+368775.4	6014566.8	2230.211	4.7198	0.05
+360094.1	6011232.9	2205.465	1.239	0.05
+366217.8	6013152.7	2310.513	3.6873	0.05
+362311	6012305.1	2287.369	-5.3138	0.05
+366617	6013923.9	2328.716	4.1414	0.05
+363307.2	6013055	2350.074	-3.5809	0.05
+359341.6	6011480.9	2188.002	3.9435	0.05
+359233.7	6010404.5	2202.372	4.0784	0.05
+367064.8	6010915.6	2176.537	-1.1443	0.05
+359468.9	6013023	2187.709	3.9948	0.05
+360683.6	6009231	2185.043	0.5588	0.05
+363405.8	6008464.9	2171.398	-14.0378	0.05
+363040	6007708	2181.436	-14.3858	0.05
+363424.3	6005859.8	2188.994	-16.3174	0.05
+361929.4	6007570.8	2196.007	-11.6554	0.05
+361322.2	6010389.6	2161.111	-2.8714	0.05
+360906.5	6005685.1	2192.528	-9.5642	0.05
+360081.2	6004739.2	2201.676	-2.2541	0.05
+358205.6	6003745.7	2328.486	-0.4587	0.05
+364459.8	6010921.5	2168.886	-5.2441	0.05
+366060.9	6004171	2184.777	4.9041	0.05
+366233.8	6004029.9	2186.003	5.5234	0.05
+359496.7	6014887.5	2161.4	3.8769	0.05
+359605.9	6013750.3	2154.978	4.0254	0.05
+359843.8	6013213.6	2185.253	3.8105	0.05
+359920.8	6012727.2	2203.143	2.8361	0.05
+360076.3	6012249.9	2220.439	2.5646	0.05
+361136.6	6011684.2	2297.049	-1.6368	0.05
+361574.2	6011880	2309.14	-4.2936	0.05
+362055.2	6012061.7	2303.044	-5.4453	0.05
+362626.4	6012780.5	2289.942	-3.3805	0.05
+363018.5	6012977.8	2325.43	-3.2894	0.05
+363481	6013343.1	2355.786	-1.8359	0.05
+363922.9	6013534.8	2357.851	-0.8487	0.05
+364482.5	6012914.5	2363.058	-1.0972	0.05
+364657	6012451.6	2351.115	-1.9203	0.05
+364923.7	6012810.5	2359.124	-0.8584	0.05
+365101	6013277.8	2358.159	0.9171	0.05
+365611.6	6013387	2346.855	2.0871	0.05
+366173.8	6013631.3	2317.519	3.6341	0.05
+366856.5	6014278.4	2312.226	3.8785	0.05
+367419	6014362.1	2298.255	4.2083	0.05
+367921.7	6014372	2261.726	3.7659	0.05
+368449	6014348.3	2244.935	4.3701	0.05
+369152.3	6014957.4	2201.091	4.9783	0.05
+369658.7	6014931.8	2201.995	5.4784	0.05
+359000.2	6012880.1	2180.407	4.2008	0.05
+358883.4	6012462.3	2177.161	4.5351	0.05
+358509.1	6012213.2	2167.542	3.8742	0.05
+358936.7	6011715.8	2184.28	3.8015	0.05
+358893.9	6011249.6	2215.462	3.7045	0.05
+358592.7	6010921.6	2213.598	4.1896	0.05
+358236.2	6010625.3	2228.342	4.1446	0.05
+357746.9	6010694.4	2268.556	3.6924	0.05
+357425.2	6011060.7	2307.562	2.995	0.05
+358422	6011412.2	2303.356	2.5713	0.05
+359364.9	6011056.6	2167.662	2.8298	0.05
+359155.3	6009880.9	2166.452	4.1893	0.05
+359539.8	6009476.7	2160.818	2.372	0.05
+360042.9	6009469.2	2165.836	2.2404	0.05
+360978.5	6009611.5	2160.491	-1.5732	0.05
+361211	6009171	2161.913	-2.9557	0.05
+361535.1	6008805.7	2162.689	-6.1186	0.05
+361675.6	6008330.5	2161.829	-8.6285	0.05
+358654.2	6009829.2	2201.714	4.2313	0.05
+358353.9	6009426.9	2297.539	3.871	0.05
+357953.9	6009057.5	2333.486	4.2597	0.05
+357441.9	6008757.3	2449.519	3.9877	0.05
+357192.9	6008325.9	2470.826	3.0748	0.05
+356836	6008077.3	2500.811	2.3348	0.05
+361367	6006269.2	2186.354	-11.6109	0.05
+360954.7	6005975.3	2201.936	-9.8473	0.05
+360545.3	6005186.9	2192.131	-5.5546	0.05
+359676	6004399.1	2216.005	-0.9334	0.05
+359183.9	6004206.2	2223.183	-0.7048	0.05
+358690	6004317.5	2240.749	-0.2499	0.05
+358214.4	6004122.5	2267.646	-1.0172	0.05
+357479.5	6004360.7	2273.342	1.1004	0.05
+357307.5	6005061.1	2257.032	0.7126	0.05
+357769.1	6005260.7	2257.991	0.05527	0.05
+358275.3	6005377.1	2253.481	-0.0448	0.05
+358990.1	6005582	2225.082	-0.612	0.05
+359426.2	6005849.7	2208.702	-2.5411	0.05
+359927.7	6005832.1	2206.934	-5.0677	0.05
+361670.5	6006416.3	2159.826	-13.3736	0.05
+368955.9	6013437	2160.188	4.7008	0.05
+368980.9	6012858.5	2160.165	2.9635	0.05
+368957.5	6012163.5	2166.898	1.7956	0.05
+368781.5	6011610.3	2169.93	1.2387	0.05
+368549.1	6011021.2	2161.041	0.4767	0.05
+368328.4	6010552	2162.389	0.3942	0.05
+368138.5	6009929.5	2160.917	0.4634	0.05
+368321.3	6009170.3	2161.018	1.6791	0.05
+368158.2	6008636.9	2160.607	-0.2096	0.05
+367855.4	6008233.3	2161.901	-1.0353	0.05
+367681	6007757.2	2161.971	-0.2368	0.05
+367535.1	6007271.6	2161.757	-0.3759	0.05
+367428.4	6006770.9	2161.644	1.899	0.05
+367265	6006275.2	2161.193	3.7072	0.05
+367045.1	6005800.6	2161.079	5.077	0.05
+366887.1	6005297.6	2168.517	6.0607	0.05
+366924.6	6004781.9	2160.798	4.3736	0.05
+367488.6	6004465.7	2160.969	4.3784	0.05
+368001	6007243.3	2186.598	2.2492	0.05
+368505.3	6007066.3	2247.928	2.8166	0.05
+368968.9	6006886.5	2294.719	5.3419	0.05
+369446.4	6006797.9	2340.813	5.8692	0.05
+369938.4	6006628.9	2452.104	4.1879	0.05
+370233.1	6006504.2	2478.823	3.9529	0.05
+364872.4	6003989.7	2162.273	0.8152	0.05
+361581.9	6006914.1	2166.497	-12.7156	0.05
+361911.1	6005892.5	2161.213	-14.397	0.05
+362432	6006128.9	2161.2	-17.4116	0.05
+362824	6006286.5	2162.131	-18.6309	0.05
+363294.6	6006160.3	2161.08	-18.1361	0.05
+367621.8	6013497.1	2169.281	4.9482	0.05
+367929.7	6013094.2	2168.942	3.9368	0.05
+367794.9	6012263.7	2167.65	1.3244	0.05
+367223.5	6011536.7	2168.687	-0.3412	0.05
+366630	6012199.2	2168.907	1.6501	0.05
+365904	6011713.2	2165.875	-1.2276	0.05
+365936.2	6011209.1	2167.131	-1.8156	0.05
+365731.2	6010915	2166.87	-3.2852	0.05
+365559.7	6011326.6	2167.944	-3.0949	0.05
+365012.1	6011050	2167.057	-4.1952	0.05
+364447.6	6011501.2	2165.267	-3.4829	0.05
+364241	6011986.1	2164.931	-3.7844	0.05
+364058.3	6012445.2	2165.348	-3.2654	0.05
+363711	6012686.4	2165.028	-3.0893	0.05
+363266.2	6012412.6	2165.095	-3.9063	0.05
+362835.3	6012198.5	2165.437	-4.9381	0.05
+362636.6	6011642.8	2167.954	-7.2193	0.05
+361975.2	6011005.6	2169.75	-5.1045	0.05
+360885.9	6012098.7	2332.313	-0.9737	0.05
+360778.7	6012622.7	2334.431	-0.9087	0.05
+360455.8	6012993.8	2340.932	0.2246	0.05
+360454.1	6013514.4	2417.598	-0.8988	0.05
+360550.5	6014052.3	2428.877	-0.4728	0.05
+360566.5	6014531.8	2431.259	-0.1193	0.05
+360973.1	6014173.3	2431.723	-0.0183	0.05
+361486.2	6014206.6	2463.004	-1.007	0.05
+361920.7	6014382.1	2454.613	-0.8176	0.05
+362414.5	6014385.4	2450.971	-1.4539	0.05
+364030.6	6003362.7	2230.716	-0.1599	0.05
+363663.9	6003017.9	2254.717	-0.0891	0.05
+363234.1	6002874.9	2270.788	-1.348	0.05
+363092	6002514.6	2342.944	0.5039	0.05
+362773.8	6002066.6	2449.163	-0.6829	0.05
+362517.3	6007599.3	2169.206	-13.8463	0.05
+363537	6008139.4	2164.473	-15.6983	0.05
+362791.7	6008156.1	2169.742	-12.7892	0.05
+363388.1	6009012.9	2164.473	-12.4353	0.05
+366321.5	6003653.7	2202.612	5.1697	0.05
+366492.5	6003153.4	2185.419	7.0861	0.05
+366487.4	6002625.3	2175.218	8.8288	0.05
+366462.9	6002138.2	2182.963	9.5639	0.05
+366508.5	6001657.7	2183.867	9.917	0.05
+366728.8	6001227.8	2186.336	9.1002	0.05
+366911.7	6000761.2	2194.478	7.9827	0.05
+367200.5	6000531.6	2193.351	7.493	0.05
+361506	6010860.6	2164.254	-3.2873	0.05
+360811.4	6010798.2	2166.447	0.0936	0.05
+367800.2	6003356.3	2180.167	3.4612	0.05
+368232.2	6003116.2	2219.28	2.9787	0.05
+368576.2	6002753.4	2243.722	1.6729	0.05
+369019.6	6002527.9	2262.227	2.4391	0.05
+360480.2	6004042.8	2252.471	-2.8087	0.05
+360908.7	6003672.5	2320.061	-2.449	0.05
+360993.5	6003060.7	2341.764	-0.9979	0.05
+361094.1	6002503.2	2353.045	-0.1978	0.05
+361352	6002090	2402.026	-0.108	0.05
+361664.8	6001615.8	2460.82	-0.2636	0.05
+362123.8	6001462.5	2490.509	0.0878	0.05
+365749.3	6004425.5	2169.68	4.4086	0.05
+363682.4	6000022.4	2865.274	-0.6574	0.05
+357107.2	6007140.4	2504.079	0.8159	0.05
+368790.4	6004470.7	2394.426	7.4507	0.05
+366071.1	6001990.8	2247.601	7.1333	0.05
+356534.1	6013529.4	2495.766	1.1395	0.05
+357297.5	6002758.7	2586.419	-0.0255	0.05
+360284.7	6000220.3	2741.396	1.077	0.05
+359283.8	6002004.4	2769.861	-2.9513	0.05
+362484.3	5999890.6	2693.787	1.5806	0.05
