+++ demo_examples/Chile_GRAV_4_Miller/LdM_mesh.mesh
@@ -1,0 +1,5 @@
+69 69 40
+355000 5999000 3000
+69*250
+69*250
+5*100 5*120 5*145 5*175 5*210 5*250 5*300 5*350
