--- docs/demo_examples/README.txt
@@ -1,2 +1,0 @@
-Examples
-========
