+++ docs/demo_examples/plot_cahn_hilliard.rst
@@ -158,7 +158,7 @@ as the time increases.
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  4.460 seconds)
+   **Total running time of the script:** ( 0 minutes  4.513 seconds)
 
 
 .. _sphx_glr_download_demo_examples_plot_cahn_hilliard.py:
