+++ docs/demo_examples/plot_dc_resistivity.rst
@@ -83,7 +83,7 @@ Basic Forward 2D DC Resistivity
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  1.002 seconds)
+   **Total running time of the script:** ( 0 minutes  0.998 seconds)
 
 
 .. _sphx_glr_download_demo_examples_plot_dc_resistivity.py:
