+++ docs/demo_examples/plot_mesh_types.rst
@@ -58,7 +58,7 @@ Here we show :code:`discretize` used to create three different types of meshes.
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  0.154 seconds)
+   **Total running time of the script:** ( 0 minutes  0.163 seconds)
 
 
 .. _sphx_glr_download_demo_examples_plot_mesh_types.py:
