+++ docs/demo_examples/plot_quadtree_hanging.rst
@@ -60,7 +60,7 @@ on an 8x8 mesh (2^3).
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  0.100 seconds)
+   **Total running time of the script:** ( 0 minutes  0.101 seconds)
 
 
 .. _sphx_glr_download_demo_examples_plot_quadtree_hanging.py:
