+++ docs/demo_examples/plot_tensors.rst
@@ -55,7 +55,7 @@ notation::
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  0.078 seconds)
+   **Total running time of the script:** ( 0 minutes  0.085 seconds)
 
 
 .. _sphx_glr_download_demo_examples_plot_tensors.py:
