+++ docs/demo_examples/sg_execution_times.rst
@@ -5,17 +5,17 @@
 
 Computation times
 =================
-**00:10.944** total execution time for **demo_examples** files:
+**00:11.422** total execution time for **demo_examples** files:
 
-- **00:04.460**: :ref:`sphx_glr_demo_examples_plot_cahn_hilliard.py` (``plot_cahn_hilliard.py``)
-- **00:03.176**: :ref:`sphx_glr_demo_examples_plot_slicer_demo.py` (``plot_slicer_demo.py``)
-- **00:01.079**: :ref:`sphx_glr_demo_examples_plot_streamThickness.py` (``plot_streamThickness.py``)
-- **00:01.002**: :ref:`sphx_glr_demo_examples_plot_dc_resistivity.py` (``plot_dc_resistivity.py``)
-- **00:00.251**: :ref:`sphx_glr_demo_examples_plot_image.py` (``plot_image.py``)
-- **00:00.222**: :ref:`sphx_glr_demo_examples_plot_cyl_mirror.py` (``plot_cyl_mirror.py``)
-- **00:00.169**: :ref:`sphx_glr_demo_examples_plot_3D_cyl.py` (``plot_3D_cyl.py``)
-- **00:00.154**: :ref:`sphx_glr_demo_examples_plot_mesh_types.py` (``plot_mesh_types.py``)
-- **00:00.152**: :ref:`sphx_glr_demo_examples_plot_quadtree_divergence.py` (``plot_quadtree_divergence.py``)
-- **00:00.100**: :ref:`sphx_glr_demo_examples_plot_quadtree_hanging.py` (``plot_quadtree_hanging.py``)
-- **00:00.100**: :ref:`sphx_glr_demo_examples_plot_quadtree_refine.py` (``plot_quadtree_refine.py``)
-- **00:00.078**: :ref:`sphx_glr_demo_examples_plot_tensors.py` (``plot_tensors.py``)
+- **00:04.513**: :ref:`sphx_glr_demo_examples_plot_cahn_hilliard.py` (``plot_cahn_hilliard.py``)
+- **00:03.525**: :ref:`sphx_glr_demo_examples_plot_slicer_demo.py` (``plot_slicer_demo.py``)
+- **00:01.081**: :ref:`sphx_glr_demo_examples_plot_streamThickness.py` (``plot_streamThickness.py``)
+- **00:00.998**: :ref:`sphx_glr_demo_examples_plot_dc_resistivity.py` (``plot_dc_resistivity.py``)
+- **00:00.265**: :ref:`sphx_glr_demo_examples_plot_image.py` (``plot_image.py``)
+- **00:00.255**: :ref:`sphx_glr_demo_examples_plot_cyl_mirror.py` (``plot_cyl_mirror.py``)
+- **00:00.185**: :ref:`sphx_glr_demo_examples_plot_3D_cyl.py` (``plot_3D_cyl.py``)
+- **00:00.163**: :ref:`sphx_glr_demo_examples_plot_mesh_types.py` (``plot_mesh_types.py``)
+- **00:00.153**: :ref:`sphx_glr_demo_examples_plot_quadtree_divergence.py` (``plot_quadtree_divergence.py``)
+- **00:00.101**: :ref:`sphx_glr_demo_examples_plot_quadtree_hanging.py` (``plot_quadtree_hanging.py``)
+- **00:00.099**: :ref:`sphx_glr_demo_examples_plot_quadtree_refine.py` (``plot_quadtree_refine.py``)
+- **00:00.085**: :ref:`sphx_glr_demo_examples_plot_tensors.py` (``plot_tensors.py``)
