--- tutorial_examples/README.txt
@@ -1,4 +1,0 @@
-.. _tutorials:
-
-Tutorials
-=========
