+++ tutorial_examples/mesh_generation/tree_mesh.py
@@ -1,0 +1,7 @@
+"""
+Tree meshes
+===========
+
+Some stuff about tree meshes in the docstring
+
+"""
